move raw tests
