+++ tests_zemu/tests/standard.test.ts
@@ -415,64 +415,7 @@ describe("Standard", function () {
     },
   );
 
-  test.concurrent.each(models)("RawBytes", async function (m) {
-    const sim = new Zemu(m.path);
-    try {
-      await sim.start({ ...defaultOptions, model: m.name });
-      const app = new FilecoinApp(sim.getTransport());
 
-      // Put the app in expert mode
-      await sim.toggleBlindSigning();
-
-      // sign 1KB of data
-      const raw_bytes = Buffer.from(
-        "ab11c412ff5f6fafc466e856f67eb20ad85ef754ad1b7c5d4120ffe95dcd94bd1079f1a89a575d284422825f1aaeb099439bc60e6537e3c939a3a5f0e108d372be73d388da351c11bfc5a20a316051fcd52b4a6d003cd1eef171ba197cfbf8d245f705d65ee0c82fa74e4d3ee1f918a496a0244fb342b7ea0a836e522ba3519001866edde3207af56ad45177433ceb0290e0b55e0584b4c799a7646805d50e885e95e89209d5b223d82001be1c85c881ec6c5bd21bcfceb286c12fdc1f28feaaaa13853655c24f6ef5c640c222ba8ed161718d535786867481fb96bc1720be4b63438d72ba559cb0c72485d1fb6543bc6c684d358aa7cfc1877031600c6efb0f90e5224951205e276cbbd3876953e92a522e26d22a75b0417b2971866a839c03825df7e06de380e00ba7599c59a01165a0ac95d636cc63d09f095df058a273aa4067e9dbeeb7d28ba62519c34c485c9389a485d90f6c47698260fc43b5d2fb88794c34f129fd2861a310c74238f12cd7c84b4f8df19faf05a0756e8b5261b48ee45929f9cfc33c8cedb69029af312a544b216ea8fc33a10cd7188d58591c8a22b2ee3ab6816fe45e080c4f1733ea2a71627cbc90133cecd8eae635e0d522731ee1992a09f411a424bc48ae54cfebcdb442d34ef8e42b1cd9212fdda322baed3569437e1106b67a25d064b0d96a1150a4ea866e4849eb646574a5e3c0d4d6efca09eef7feaf540a6eda9c886d92018b2afbf64d9c077c83f23f45529f826a51b575432c6fa0c7849799c3e9ba5a0f4d71b93a12b72a9d06238c686561cd952a2a50e2c516f3fc1b60e94365dbc883a8a47a0214a6df74390c9963836e6d1099bc16da0a6caf07f0962b945ef225930bd6131fe344ff7fcac9f0181a0a24940146b03b79a3de67b92fe592183258e939685d47089e6f9228b169952aabb45f3ad369b1d557099ce97b6092f2e0bd6122c2479fed1a2427c8fd763a93587795f38a391782b0dadf857a3a8d896940c94cef4183d3ff52f26af4957736955db70d668f524285d091313ffc9b807e0502edc6fbc3f1d6e76350a0c3d78fc6cdc6ae36bd2b9dccb3b4e7734c8d91a2c883390953429fd9dd185a81bfa3ac147d86342ac3b227eff6ac0c2904596076b845a3267b1b472e8bbb429575fb280ec82718734ceb2b07e8c998b42cad224c98cc56aa5ca3a9159e8bf3604f4f56b2350befc00cca8e1a1aecb3dbb64c9536ec557204dfd3ee68ee16b641c41e75c4f97266ed4c5f78b5f8fd7ff11eb8c5db201f85b3904f13931bbead263a00e85d1086340bb4a2fb6fd139b793d4a7540b3dbf2495f7d08f8821759bde65817aa08fa1424101639fbfb6c4f91961da1372bccb127afc627d352f9d9d2faa5a9176be55274b53dc04b94174b6b7aa52955939cf14970d31e03ea60cb2cdc99e422f232a4052",
-        "hex",
-      );
-      const prefix = Buffer.from("Filecoin Sign Bytes:\n");
-      const txBlob = Buffer.concat([prefix, raw_bytes]);
-      console.log("Tx: ", txBlob.toString("hex"));
-      const d = getDigest(txBlob);
-      const print = Buffer.from(d);
-      console.log("digest: ", print.toString("hex"));
-
-      const pkResponse = await app.getAddressAndPubKey(PATH);
-      console.log(pkResponse);
-      expect(pkResponse.return_code).toEqual(0x9000);
-      expect(pkResponse.error_message).toEqual("No errors");
-
-      // do not wait here..
-      const signatureRequest = app.signRawBytes(PATH, txBlob);
-
-      await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
-      await sim.compareSnapshotsAndApprove(
-        ".",
-        `${m.prefix.toLowerCase()}-sign_raw_bytes`,
-        true,
-        0,
-        1500,
-        true,
-      );
-
-      let resp = await signatureRequest;
-      console.log(resp);
-
-      expect(resp.return_code).toEqual(0x9000);
-      expect(resp.error_message).toEqual("No errors");
-
-      // Verify signature
-      const pk = Uint8Array.from(pkResponse.compressed_pk);
-      const digest = getDigest(txBlob);
-
-      const signature = secp256k1.signatureImport(
-        Uint8Array.from(resp.signature_der),
-      );
-      const signatureOk = secp256k1.ecdsaVerify(signature, digest, pk);
-      expect(signatureOk).toEqual(true);
-    } finally {
-      await sim.close();
-    }
-  });
 
   test.concurrent.each(models)("InvokeEVM_ERC20Transfer", async function (m) {
     const sim = new Zemu(m.path);
